Update Spring and related dependencies
This commit updates Spring to the latest version which is compatible
with Java 8. Inline with this some of the dependencies have been
upgraded to

Hibernate - 5.2
SLF4J - 1.7

The configurations have also been updated to reflect the changes to the
dependencies.

While fixing the tests for the newer upgrades the Sloth based tests
where failing due to a very slow SocketChannel.connect. The tests have
been updated to better handle slow tests (instead of using a long
Thread.sleep) and the Sloth implementation is now forced to work against
the loopback interface.